WIP: splitting tests and code
