+++ script/tests/fixtures/over_66%_voting_power_change/commit.json
