+++ script/tests/fixtures/over_66%_voting_power_change/commit?height=177842.json
