+++ script/tests/fixtures/over_66%_voting_power_change/status.json
