+++ script/tests/fixtures/over_66%_voting_power_change/validators?height=177840.json
